# backend/openapi/openapi-draft.yaml
openapi: "3.0.3"
info:
  title: NewHaven API
  version: "0.1.0"
  description: Backend API for NewHaven browser-based economy simulation game
servers:
  - url: http://localhost:8088
paths:
...
  /api/v2/production/jobs/:
    get:
      tags: [Production]
      summary: List production jobs for my company
      operationId: listProductionJobs
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Production job list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductionJobListResponse'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v2/production/start/:
    post:
      tags: [Production]
      summary: Start a production job
      operationId: startProduction
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StartProductionRequest'
      responses:
        "200":
          description: Production job started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StartProductionResponse'
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v2/production/claim/{jobId}/:
    post:
      tags: [Production]
      summary: Claim produced resources from a production job
      operationId: claimProduction
      security:
        - BearerAuth: []
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Production claimed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClaimProductionResponse'
        "400":
          description: Nothing claimable or conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "404":
          description: Job not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v2/production/claimable/:
    get:
      tags: [Production]
      summary: List claimable production jobs
      operationId: listClaimableJobs
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Claimable job list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClaimableJobListResponse'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v3/resources/:
    get:
      tags: [Market]
      summary: List market-tradable resources
      operationId: listResources
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Resource list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourcesResponse'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  # --- Bonds ---
...
  /api/v3/market-ticker/{resourceId}/:
    get:
      tags: [Market]
      summary: Get market ticker for a resource
      operationId: getMarketTicker
      security:
        - BearerAuth: []
      parameters:
        - name: resourceId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Market ticker data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MarketTickerResponse'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v3/market-depth/{resourceId}/{quality}/:
    get:
      tags: [Market]
      summary: Get market depth (order book) for a resource and quality
      operationId: getMarketDepth
      security:
        - BearerAuth: []
      parameters:
        - name: resourceId
          in: path
          required: true
          schema:
            type: integer
        - name: quality
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Market depth data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MarketDepthResponse'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v3/market/{resourceId}/{quality}/:
    get:
      tags: [Market]
      summary: List market orders for a resource and quality
      operationId: listMarketOrders
      security:
        - BearerAuth: []
      parameters:
        - name: resourceId
          in: path
          required: true
          schema:
            type: integer
        - name: quality
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Market order list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MarketOrderListResponse'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        data:
          nullable: true
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object

    # --- Company ---
...
    ResourceDefinition:
      type: object
      properties:
        resourceId:
          type: integer
        name:
          type: string
        producedFrom:
          type: object
          additionalProperties:
            type: integer
        producedPerHourRaw:
          type: integer
        unitsSoldAnHour:
          type: integer
        hasEconomyModel:
          type: boolean
    ResourcesResponse:
      type: object
      properties:
        resources:
          type: array
          items:
            $ref: '#/components/schemas/ResourceDefinition'
    MarketTickerPoint:
      type: object
      properties:
        price:
          type: number
        time:
          type: string
          format: date-time
    MarketTickerResponse:
      type: object
      properties:
        resource:
          type: integer
        series:
          type: array
          items:
            $ref: '#/components/schemas/MarketTickerPoint'
    MarketDepthLevel:
      type: object
      properties:
        price:
          type: number
        quantity:
          type: integer
        qty:
          type: integer
    MarketDepthResponse:
      type: object
      properties:
        buys:
          type: array
          items:
            $ref: '#/components/schemas/MarketDepthLevel'
        sells:
          type: array
          items:
            $ref: '#/components/schemas/MarketDepthLevel'
    MarketOrderDTO:
      type: object
      properties:
        id:
          type: string
        resourceId:
          type: integer
        kind:
          type: integer
          enum: [0, 1]
        price:
          type: number
        quality:
          type: integer
        quantity:
          type: integer
        remaining:
          type: integer
        companyId:
          type: integer
        createdAt:
          type: string
          format: date-time
        status:
          type: string
    MarketOrderListResponse:
      type: object
      properties:
        orders:
          type: array
          items:
            $ref: '#/components/schemas/MarketOrderDTO'
...
    StartProductionRequest:
      type: object
      required: [building_id, resource_id, quantity]
      properties:
        building_id:
          type: string
        resource_id:
          type: integer
        quantity:
          type: integer
          minimum: 1
...
    ProductionJobDTO:
      type: object
      properties:
        id:
          type: string
        building_id:
          type: string
        resource_id:
          type: integer
        quantity:
          type: integer
        target_quantity:
          type: integer
        started_at:
          type: string
          format: date-time
        duration_seconds:
          type: number
        claimed_amount:
          type: integer
        claimable_amount:
          type: integer
        status:
          type: string
          enum: [running, ready, claimed]
...
    ProductionJobListResponse:
      type: object
      properties:
        jobs:
          type: array
          items:
            $ref: '#/components/schemas/ProductionJobDTO'
    BuildingProductionStatus:
      type: object
      properties:
        id:
          type: string
        busy:
          type: boolean
        job_id:
          type: string
    StartProductionResponse:
      type: object
      properties:
        job:
          $ref: '#/components/schemas/ProductionJobDTO'
        building:
          $ref: '#/components/schemas/BuildingProductionStatus'
    ClaimProductionResponse:
      type: object
      properties:
        job_id:
          type: string
        status:
          type: string
          enum: [running, ready, claimed]
        output:
          type: object
          additionalProperties:
            type: integer
        claimed_amount:
          type: integer
        remaining:
          type: integer
        xp:
          type: integer
        level:
          type: integer
        market_unlocked:
          type: boolean
    ClaimableJobDTO:
      type: object
      properties:
        job_id:
          type: string
        building_id:
          type: string
        resource_id:
          type: integer
        total_amount:
          type: integer
        claimed_amount:
          type: integer
        claimable_amount:
          type: integer
    ClaimableJobListResponse:
      type: object
      properties:
        jobs:
          type: array
          items:
            $ref: '#/components/schemas/ClaimableJobDTO'

    # --- Production Queue ---
...